requirement.txt
streamlit
python-dotenv
speechrecognition
pypdf
requests
keyboard
pyaudio
